logalloc: add on-stall memory reclaim diagnostics
Beef up the existing `reclaim_timer` logic:

* Use it for debug logging and/or when a stall is detected
* Tweak log level according to conditions
* Log compaction/migration/eviction/allocation stats
* Log even when exiting "early"

Note that time measurement in the stall case is imprecise (we don't
want to call `clock::now()` before we know there is a stall) -- so
elapsed time before the first detected stall is assumed to be the
configured stall detector threshold.

Object-level migration and eviction stats are not collected yet (it
started to become something of a rat-hole and I ejected myself from it
for now).

Refs #4186.

Signed-off-by: Michael Livshin <michael.livshin@scylladb.com>