build: allow setting SCYLLA_RELEASE from outside
Current logic that derives the SCYLLA_RELEASE value has those
assumptions:

* The sources being built include Git metadata.

* Using the current time in SCYLLA_RELEASE is a good idea.

* Everything that is interesting in the worktree is committed to Git
  prior to build.

The above assumptions are either irrelevant or problematic in pure
build environments (such as sandboxes set up by `nix-build`):

* Pure builds use cleaned-up sources with all timestamps reset to Unix
  time 0, for reproducibility.  Those cleaned-up sources are saved (in
  the Nix store, for example) and content-hashed, so leaving
  the (possibly huge) Git metadata increases the time to copy the
  sources and wastes disk space.

* Pure builds run in a sandbox where time is, likewise, reset to Unix
  time 0, so current time is not useful information.

Now, the only build step that uses Git metadata in the first place is
the SCYLLA_RELEASE value derivation logic.  So, essentially, answering
the question "is Git metadata needed to build Scylla" is a matter of
definition, and is up to us.  If we elect to ignore Git metadata and
current time, we can derive SCYLLA_RELEASE value from cleaned-up
sources, regardless of the way those sources were arrived at.

This change makes it possible to skip the derivation of SCYLLA_RELEASE
value from Git metadata and time by setting SCYLLA_RELEASE in the
environment.

Signed-off-by: Michael Livshin <michael.livshin@scylladb.com>